pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "social-media-ai"
version = "0.1.0"
description = "AI-powered Social Media Management Platform"
readme = "README.md"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Communications :: Social Media",
]
requires-python = ">=3.10"
dependencies = [
    # Web Framework
    "fastapi>=0.95.0",
    "uvicorn>=0.21.0",
    "python-multipart>=0.0.6",
    
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.10.0",
    "psycopg2-binary>=2.9.6",
    
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # CLI
    "typer[all]>=0.9.0",
    "rich>=13.0.0",

    
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
cli = "cli.main:app"

[tool.setuptools]
packages = ["app", "cli"]
package-dir = {"" = "."}

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"
addopts = "-v --cov=app --cov-report=term-missing"

[project.urls]
Homepage = "https://github.com/yourusername/social-media-ai"
Documentation = "https://github.com/yourusername/social-media-ai#readme"
Issues = "https://github.com/yourusername/social-media-ai/issues"
